feat: text body preprocessing
