--- server/api/script1.py
@@ -1,1 +1,0 @@
-print('hello World')
